ohlcv/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
dateparser==1.1.1
finnhub-python==2.4.14
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
psycopg2==2.9.3
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
tzdata==2022.4
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
websockets==10.3
yarl==1.8.1
